Delete BoundingBoxExtensions
