+++ Source/HelixToolkit.SharpDX/Utilities/Octrees/MeshGeometryOctree.cs
@@ -46,7 +46,7 @@ public class MeshGeometryOctree
     {
         Positions = positions;
         Indices = indices;
-        Bound = BoundingBoxExtensions.FromPoints(positions);
+        Bound = BoundingBoxHelper.FromPoints(positions);
         Objects = indices is null ? null : new List<KeyValuePair<int, BoundingBox>>(indices.Count / 3);
         // Construct triangle index and its bounding box KeyValuePair
         if (indices is not null)
